+++ packages/davatar-storybook/src/App.tsx
@@ -1,3 +1,4 @@
+import * as React from "react";
 import { useState } from 'react';
 import './App.css';
