+++ packages/davatar-storybook/src/stories/ScribeApp.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import * as React from "react";
 import type { ComponentStory, ComponentMeta } from "@storybook/react";
 import { Scribe } from "davatar-ui";
