+++ packages/davatar-storybook/tsconfig.json
@@ -14,7 +14,7 @@
     "moduleResolution": "Node",
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "jsx": "react-jsx",
+    "jsx": "react",
     "typeRoots": [
       "./types",
       "./node_modules/@types",
